+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### Unreleased
+
+**New features**
+
+- Support non-ARM polling by default (azure-core 1.4.0)
+
 ### 2020-04-09 - 5.0.0-dev.20200409.1
 Modelerfour version: 4.12.301